fix: 아바타 id 값 Long 형으로 변경

+++ app/src/main/java/com/into/websoso/data/model/AvatarEntity.kt
@@ -1,7 +1,7 @@
 package com.into.websoso.data.model
 
 data class AvatarEntity(
-    val avatarId: Int,
+    val avatarId: Long,
     val avatarName: String,
     val avatarLine: String,
     val avatarProfileImage: String,
